Start pim functionsfex
