+++ dev/others/main-azureAuth.go
